feat: add theme for signature pad #3 - add useSignaturePad Hook
