+++ src/components/SignaturePad.tsx
@@ -1,136 +1,58 @@
-import React, { useRef, useEffect, useState } from "react";
+import React from "react";
 import Icons from "./Icons";
 import { type SignaturePadProps } from "../types";
+import useSignaturePad from "../hooks/useSignaturePad";
+import { THEME } from "../utils/theme";
 
 const SignaturePad: React.FC<SignaturePadProps> = ({
-  width,
-  height,
-  penColor = "black",
+  width = 300,
+  height = 150,
   onCopy,
   onDownload,
   onClear,
+  theme = "light",
   download = false,
   downloadFilename = "signature",
   downloadFormat = "image/png",
-  padStyleClassName = {},
+  padStyles = {},
   iconColor = {},
 }) => {
-  const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const [isDrawing, setIsDrawing] = useState(false);
-  const [isClear, SetIsClear] = useState(true);
+  const {
+    canvasRef,
+    isClear,
+    clearCanvas,
+    draw,
+    handleCopy,
+    handleDownload,
+    startDrawing,
+    stopDrawing,
+  } = useSignaturePad(
+    width,
+    height,
+    theme,
+    downloadFilename,
+    downloadFormat,
+    onCopy,
+    onDownload,
+    onClear
+  );
 
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    const context = canvas?.getContext("2d");
-    if (context) {
-      context.fillStyle = `${
-        downloadFormat === "image/png" ? "transparent" : "white"
-      }`;
-      context.fillRect(0, 0, width, height);
-      context.lineCap = "round";
-      context.strokeStyle = penColor;
-      context.lineWidth = 2;
-    }
-  }, [penColor]);
-
-  const clearCanvas = () => {
-    if (!isClear) {
-      const canvas = canvasRef.current;
-      const ctx = canvas?.getContext("2d");
-
-      if (canvas && ctx) {
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        ctx.fillStyle = `${
-          downloadFormat === "image/png" ? "transparent" : "white"
-        }`;
-        ctx.fillRect(0, 0, canvas.width, canvas.height);
-        onClear && onClear();
-        SetIsClear(true);
-      }
-    }
-  };
-
-  const startDrawing = (
-    event:
-      | React.MouseEvent<HTMLCanvasElement>
-      | React.TouchEvent<HTMLCanvasElement>
-  ) => {
-    const canvas = canvasRef.current;
-    const context = canvas?.getContext("2d");
-
-    if (!canvas || !context) return;
-
-    setIsDrawing(true);
-    SetIsClear(false);
-    const { offsetX, offsetY } = getPosition(event);
-
-    context.beginPath();
-    context.moveTo(offsetX, offsetY);
-  };
-
-  const draw = (
-    event:
-      | React.MouseEvent<HTMLCanvasElement>
-      | React.TouchEvent<HTMLCanvasElement>
-  ) => {
-    const canvas = canvasRef.current;
-    const context = canvas?.getContext("2d");
-
-    if (!isDrawing || !canvas || !context) return;
-
-    const { offsetX, offsetY } = getPosition(event);
-
-    context.lineTo(offsetX, offsetY);
-    context.stroke();
-  };
-
-  const stopDrawing = () => {
-    setIsDrawing(false);
-  };
-
-  const getPosition = (
-    event:
-      | React.MouseEvent<HTMLCanvasElement>
-      | React.TouchEvent<HTMLCanvasElement>
-  ) => {
-    const canvas = canvasRef.current;
-    const rect = canvas?.getBoundingClientRect();
-
-    if (event.type.includes("touch")) {
-      const touch = (event as React.TouchEvent).touches[0];
-      return {
-        offsetX: touch.clientX - (rect?.left ?? 0),
-        offsetY: touch.clientY - (rect?.top ?? 0),
-      };
-    }
-
-    const { offsetX, offsetY } = event.nativeEvent as MouseEvent;
-    return { offsetX, offsetY };
-  };
-
-  const handleCopy = () => {
-    if (canvasRef.current && !isClear) {
-      const dataURL = canvasRef.current.toDataURL();
-      onCopy && onCopy(dataURL);
-    }
-  };
-
-  const handleDownload = () => {
-    if (canvasRef.current && !isClear) {
-      const dataURL = canvasRef.current.toDataURL(downloadFormat);
-      const fileName = `${downloadFilename}.${downloadFormat.split("/")[1]}`;
-      const link = document.createElement("a");
-      link.href = dataURL;
-      link.download = fileName;
-      link.click();
-      onDownload && onDownload(dataURL);
-    }
+  const defaultPadStyles: React.CSSProperties = {
+    position: "relative",
+    width: width,
+    height: height,
+    background: THEME.backGroundColor[theme],
   };
 
   return (
-    <div style={{ position: "relative", width: width, height: height }}>
+    <div
+      role="img"
+      style={{
+        ...padStyles,
+        ...defaultPadStyles,
+      }}
+    >
       <canvas
-        role="img"
         ref={canvasRef}
         width={width}
         height={height}
@@ -141,7 +63,6 @@ const SignaturePad: React.FC<SignaturePadProps> = ({
         onTouchStart={startDrawing}
         onTouchMove={draw}
         onTouchEnd={stopDrawing}
-        style={padStyleClassName}
       ></canvas>
       <div
         onClick={clearCanvas}
@@ -177,6 +98,7 @@ const SignaturePad: React.FC<SignaturePadProps> = ({
         {download && (
           <div
             onClick={handleDownload}
+            role="icon-download"
             style={{ cursor: !isClear ? "pointer" : "auto" }}
           >
             <Icons
