+++ src/types/index.ts
@@ -5,15 +5,27 @@ export interface IconColor {
 }
 
 export interface SignaturePadProps {
-  width: number;
-  height: number;
-  penColor?: string;
+  width?: number;
+  height?: number;
+  theme?: "light" | "dark";
   onCopy?: (dataURL: string) => void;
   onDownload?: (dataURL: string) => void;
   onClear?: () => void;
   download?: boolean;
   downloadFilename?: string;
   downloadFormat?: "image/png" | "image/jpeg";
-  padStyleClassName?: React.CSSProperties;
+  padStyles?: React.CSSProperties;
   iconColor?: IconColor;
+}
+
+type BlackOrWhite = "white" | "black";
+export interface THEME {
+  penColor: {
+    light: BlackOrWhite;
+    dark: BlackOrWhite;
+  };
+  backGroundColor: {
+    light: BlackOrWhite;
+    dark: BlackOrWhite;
+  };
 }
