+++ src/utils/theme.ts
@@ -1,0 +1,14 @@
+import { THEME } from "../types";
+
+const THEME: THEME = {
+  penColor: {
+    light: "black",
+    dark: "white",
+  },
+  backGroundColor: {
+    light: "white",
+    dark: "black",
+  },
+};
+
+export { THEME };
